Update assert to print unsigned size_t

+++ libyaclib/assert.c
@@ -4,7 +4,6 @@
 
 _Noreturn void __assert_failure(const char *file, size_t line, const char *func) {
   // TODO: Output to stderr, not stdout.
-  // TODO: Unsigned line number.
-  printf("Assert failure:\n  File: %s\n  Func: %s\n  Line: %d\n", file, func, (int)line);
+  printf("Assert failure:\n  File: %s\n  Func: %s\n  Line: %zu\n", file, func, line);
   abort();
 }
